requirements.txt
# Minimum packages for azure functions
azure-functions
# Function Packages
azure-identity
azure-keyvault-secrets
pandas
sqlalchemy
psycopg2
